terraform_eks_with_ci-cd/requirements.txt
Flask
flask-mysqldb
boto3
aws_secretsmanager_caching
mysql-connector

